Fixed error in continue_statement() where it crashed on script resume: 	* Replace setjmp(), longjmp() with call to ExecuteFile(). 	* Updated ExecuteFile() to allow for resumption of script execution using NULL parameter.
Fixed error in ExecuteFile() where it did not update FileBufferPosition.
Fixed error in SetLastError() where the variable was passed to UpdateVariable, instead of a pointer.
Fixed error in wend_statement() where 0 was not returned.
Fixed error in while_statement() where the buffer position was not save before executing the loop and restored at the end of the loop.
Fixed error in ResizeArray() where it did not handle user-defined types.

Added ability to resize arrays:
	* Added resize_statement() function.
	* Added resize to list of statements.
	* Added function prototype to dofile.h.

Fixed error in ResizeArray() where it returned -1 on error instead of error number.
Fixed error in assignment code where it did not check if variable was new in string assignment.
Fixed error in SubstituteVariables() where it did not check if variable is string slice.
Fixed error in print_statement() where it did not check if comma was a subscript value separator.