Changed CallFunction() return type from double to int. Changed ReturnFromFunction() return type from int to void.
Fixed error in CallFunction() where it did not check if the number of arguments was correct:
	* Added code to CallFunction to check if number of arguments is correct
	* Added error TOO_MANY_ARGUMENTS_IN_FUNCTION_CALL to errors.h
	* Added error message "Too many arguments" to list of error messages
	* Changed code to assign values to function arguments to assign by looping over the parameter values rather than over the parameter names.

Fixed error in declare_statement() where it used a fixed string rather than the default type from the variable types array:
	* Added extern declaration for vartypenames
	* Changed code in declare_statement() where it used a fixed string rather than the default type from the variable types array

Changed code in return_statement() to test if parameter is number to use IsNumber().
Fixed error in IsValidExpression() where it did not check if last argument is an operator.