Renamed removenewline() to RemoveNewline(): 	* Rename function removenewline() to RemoveNewline() 	* Renamed calls to removenewline() to RemoveNewline() in ExecuteLine(), for_statement(), InteractiveMode(), FindEndOfFunction()
Fixed error in PopSaveInformation() where it did not return -1 if stack is empty.
Fixed error in PopSaveInformation() where it did not check if stack was empty before getting line buffer address and line number.
Fixed error in IsValidExpression() where it did not check if two condition operators were together.

Fixed error where condition tokens and power operators were not tokenized correctly:
	* Updated TokenizeLine() to treat <=, >=, ** operators as a single token
	* Updated EvaluateExpression(), EvaluateSingleCondition() to use modified operator tokens

Fixed error in EvaluateExpression() where it did not use the correct token index for arguments to pow().
Fixed error in EvaluateExpression() where it did not evaluate expressions with an odd number of tokens correctly.
Fixed error in EvaluateExpression() where it did not evaluate expressions with multiple and different types of operators correctly.