Fixed error in FreeVariablesList() where it did not check if string variable pointer in variable values were not NULL. Fixed error in FreeVariablesList() where it did not free variable values. Fixed error in FreeVariablesList() where it did not free intial parameters.
Fixed error in InitializeMainFunction() where it did not check if DeclareFunction() returned -1:
	* Changed IntializeMainFunction() return type from void to int.

Fixed error in DeclareFunction() where ) was added as a variable.
Fixed missed potential buffer overflow error in main().
Fixed error in main() where junk data was in args in no arguments were defined.
Fixed error in IntializeMainFunction() where it uncessarily passed the command-line arguments to DeclareFunction() and caused them to be interpreted as a variable.
Fixed error in main() where NULL was passed as a parameter to sigaction().